+++ sites/partners/src/listings/PaperListingForm/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useContext } from "react"
+import React, { useState, useContext, useEffect } from "react"
 import { useRouter } from "next/router"
 import {
   AuthContext,
@@ -13,11 +13,11 @@ import {
 } from "@bloom-housing/ui-components"
 import { useForm, FormProvider } from "react-hook-form"
 import {
-  ListingCreate,
   ListingStatus,
-  ListingUpdate,
   CSVFormattingType,
   CountyCode,
+  Unit,
+  Listing,
 } from "@bloom-housing/backend-core/types"
 import { YesNoAnswer } from "../../applications/PaperApplicationForm/FormTypes"
 
@@ -27,16 +27,18 @@ import AdditionalDetails from "./sections/AdditionalDetails"
 import AdditionalEligibility from "./sections/AdditionalEligibility"
 import LeasingAgent from "./sections/LeasingAgent"
 import AdditionalFees from "./sections/AdditionalFees"
+import Units from "./sections/Units"
+import { stringToBoolean, stringToNumber } from "../../../lib/helpers"
+import { useAmiChartList } from "../../../lib/hooks"
 import BuildingDetails from "./sections/BuildingDetails"
 import ListingIntro from "./sections/ListingIntro"
 import BuildingFeatures from "./sections/BuildingFeatures"
 import RankingsAndResults from "./sections/RankingsAndResults"
 
-export type FormListing = ListingCreate &
-  ListingUpdate & {
-    waitlistOpenQuestion?: YesNoAnswer
-    waitlistSizeQuestion?: YesNoAnswer
-  }
+export type FormListing = Listing & {
+  waitlistOpenQuestion?: YesNoAnswer
+  waitlistSizeQuestion?: YesNoAnswer
+}
 
 type ListingFormProps = {
   listing?: FormListing
@@ -46,6 +48,9 @@ type ListingFormProps = {
 type AlertErrorType = "api" | "form"
 
 const defaultAddress = {
+  id: undefined,
+  createdAt: undefined,
+  updatedAt: undefined,
   city: "",
   state: "",
   street: "",
@@ -53,6 +58,9 @@ const defaultAddress = {
 }
 
 const defaults: FormListing = {
+  id: undefined,
+  createdAt: undefined,
+  updatedAt: undefined,
   applicationAddress: defaultAddress,
   applicationDueDate: new Date(),
   applicationFee: "0",
@@ -107,6 +115,27 @@ const defaults: FormListing = {
   unitAmenities: "",
   servicesOffered: "",
   yearBuilt: 2021,
+  urlSlug: undefined,
+  showWaitlist: false,
+  unitsSummarized: {
+    unitTypes: [],
+    reservedTypes: [],
+    priorityTypes: [],
+    amiPercentages: [],
+    byUnitTypeAndRent: [],
+    byUnitType: [],
+    byNonReservedUnitType: [],
+    byReservedType: [],
+    byAMI: [],
+    hmi: {
+      columns: [],
+      rows: [],
+    },
+  },
+}
+
+export type TempUnit = Unit & {
+  tempId?: number
 }
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -120,8 +149,23 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
 
   const { listingsService } = useContext(AuthContext)
 
+  /**
+   * fetch options
+   */
+  const { data: amiCharts = [] } = useAmiChartList()
   const [alert, setAlert] = useState<AlertErrorType | null>(null)
   const [loading, setLoading] = useState<boolean>(false)
+  const [units, setUnits] = useState<TempUnit[]>([])
+
+  useEffect(() => {
+    if (listing?.units) {
+      const tempUnits = listing.units.map((unit, i) => ({
+        ...unit,
+        tempId: i + 1,
+      }))
+      setUnits(tempUnits)
+    }
+  }, [listing, setUnits])
 
   // eslint-disable-next-line @typescript-eslint/unbound-method
   const { handleSubmit, clearErrors, reset, trigger, getValues } = formMethods
@@ -148,8 +192,50 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
   const formatFormData = (data: FormListing) => {
     const showWaitlistNumber =
       data.waitlistOpenQuestion === YesNoAnswer.Yes && data.waitlistSizeQuestion === YesNoAnswer.Yes
+    units.forEach((unit) => {
+      switch (unit.unitType) {
+        case "threeBdrm":
+          unit.numBedrooms = 3
+          break
+        case "twoBdrm":
+          unit.numBedrooms = 2
+          break
+        case "oneBdrm":
+          unit.numBedrooms = 1
+          break
+        default:
+          unit.numBedrooms = null
+      }
+
+      unit.floor = stringToNumber(unit.floor)
+      unit.maxOccupancy = stringToNumber(unit.maxOccupancy)
+      unit.minOccupancy = stringToNumber(unit.minOccupancy)
+      unit.numBathrooms = stringToNumber(unit.numBathrooms)
+
+      if (unit.sqFeet.length === 0) {
+        delete unit.sqFeet
+      }
+
+      if (unit.amiChart?.length) {
+        const chart = amiCharts.find((chart) => chart.id === unit.amiChart)
+        unit.amiChart = chart
+      } else {
+        delete unit.amiChart
+      }
+
+      if (unit.id === undefined) {
+        unit.id = ""
+        delete unit.updatedAt
+        delete unit.createdAt
+      }
+
+      delete unit.tempId
+    })
+
     return {
       ...data,
+      disableUnitsAccordion: stringToBoolean(data.disableUnitsAccordion),
+      units: units,
       isWaitlistOpen: data.waitlistOpenQuestion === YesNoAnswer.Yes,
       yearBuilt: data.yearBuilt ? Number(data.yearBuilt) : null,
       waitlistCurrentSize:
@@ -245,6 +331,12 @@ const ListingForm = ({ listing, editMode }: ListingFormProps) => {
                     <ListingIntro />
                     <FormListingData />
                     <BuildingDetails />
+                    <Units
+                      units={units}
+                      setUnits={setUnits}
+                      amiCharts={amiCharts}
+                      disableUnitsAccordion={listing?.disableUnitsAccordion}
+                    />
                     <AdditionalFees />
                     <BuildingFeatures />
                     <AdditionalEligibility />